pyproject.toml
[project]
name = "photostimulator"
version = "0.1.0"
description = ""
authors = [
    {name = "yel-gar"}
]
requires-python = ">=3.11,<3.15"
dependencies = [
    "pyserial (>=3.5,<4.0)",
    "pyside6 (>=6.11.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
